Move author to a copyright attribution
The original code here was contributed via an emailed patch in
2006 by @knawrocke:

https://sourceforge.net/p/jruby/mailman/message/9805929/

At the time we were a little loose about attribution, and the only
reference to his work was via the email list posts and the author
tag. This commit adds a copyright line with appropriate name and
email and will include metadata for co-authorship.

Co-authored-by: Kelly Nawrocke <knawrocke@gmail.com>